move search results cards to screens dir
